requirements.txt
# Python dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
redis==5.0.1
google-generativeai==0.3.1
python-multipart==0.0.6

# Additional utilities
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
flake8==6.1.0
mypy==1.7.1

# Optional: Celery for background tasks
celery==5.3.4
